Re-ordering of HDR options to make HLG default + clean-up in ini explanation

+++ cfg.h
@@ -80,9 +80,9 @@ typedef struct {
 	uint8_t vrr_vesa_framerate;
 	uint16_t video_off;
 	uint8_t disable_autofire;
-	uint16_t video_brightness;
-	uint16_t video_contrast;
-	uint16_t video_saturation;
+	uint8_t video_brightness;
+	uint8_t video_contrast;
+	uint8_t video_saturation;
 	uint16_t video_hue;
 	char video_gain_offset[256];
 	uint8_t hdr;
